refactor(badge): move badge interface to the component file
